core/common: Move list and map to common
