+++ src/common/CMakeLists.txt
@@ -4,8 +4,12 @@ add_library(common STATIC
     error.h
     hash.h
     hash.c
+    list.c
+    list.h
     logger.c
     logger.h
+    map.c
+    map.h
     string.c
     string.h
     util.c
