+++ src/common/list.c
