+++ src/common/map.h
